% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoRdseq.R
\name{recoRdseq.DE}
\alias{recoRdseq.DE}
\title{Find differentially expressed genes using Record-seq data}
\usage{
recoRdseq.DE(data, design, designFormula, tool = "DESeq2")
}
\arguments{
\item{data}{input counts matrix as data frame; rows = genes, cols = samples}

\item{design}{input design matrix as data frame; rows = samples, cols = design factors}

\item{designFormula}{any permitted combination of design factors in the design matrix, starting with "~", eg: "~Treatment+Time"}

\item{tool}{package for Differential expression. can be 'DESeq2', 'edgeR' or 'baySeq'}
}
\value{
data frame of genes with differential expression information
}
\description{
This function takes in a counts matrix, a design matrix, (optional) a designFormula and a differential expression (DE) tool.  
It outputs a data frame with the logFC and p-adj value for differential expression for each gene generated by the DE tool.
}
