% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoRdseq.R
\name{recoRdseq.filterDEG}
\alias{recoRdseq.filterDEG}
\title{Filter top DE genes based on output of recoRdseq.DE function}
\usage{
recoRdseq.filterDEG(genesDE, p = 0.05, n = NULL)
}
\arguments{
\item{genesDE}{data frame of genes with DE values generated by the recoRdseq.DE function}

\item{p}{maximum p-adj value}

\item{n}{maximum number of output genes (overrides p)}
}
\value{
vector of DE genes
}
\description{
This function takes in the data frame generated by the recoRdseq.DE function, and filters genes based on p-adjusted value
}
