% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoRdseqAnalysis.R
\name{recoRdseq.preprocess}
\alias{recoRdseq.preprocess}
\title{Preprocess data for Record-seq analysis}
\usage{
recoRdseq.preprocess(data, design, totalCounts, minCountsPerSample = 1000)
}
\arguments{
\item{data}{input counts matrix as data frame; rows = genes, cols = samples}

\item{design}{input design matrix as data frame; rows = samples, cols = design factors}

\item{totalCounts}{input cumulative counts matrix as data frame; rows = samples, cols = type of counts}

\item{minCountsPerSample}{minimum cumulative counts per sample}
}
\value{
list of pre-processed files
}
\description{
This function takes in a counts matrix, a design matrix, and (optional) a matrix of cumulative counts per sample.  
It filters and sorts samples in the counts matrices based on the design matrix. Further, it removes samples with a 
cumulative count below minCountsPerSample, and filters lowly expressed genes in the lowest quartile. It outputs a 
list, where the first element is the pre-processed counts matrix, second element is the design matrix, and third (optional)
is the processed cumulative counts matrix.
}
