% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoRdseq.R
\name{recoRdseq.transform}
\alias{recoRdseq.transform}
\title{Normalize and transform Record-seq counts data}
\usage{
recoRdseq.transform(data, design, transformation = "vst")
}
\arguments{
\item{data}{input counts matrix as data frame; rows = genes, cols = samples}

\item{design}{input design matrix as data frame; rows = samples, cols = design factors}

\item{transformation}{"log2" (general log2-transformation), "rlog" (regularized log from DESeq2), "vst" (variance stabilizing transform from DESeq2), "tmm" (trimmed mean of m normalization followed by log2 transform), "thinCounts" (downsampling followed by log2 transform)}
}
\value{
transformed counts matrix
}
\description{
This function takes in a counts matrix, a design matrix, and a transformation.  
It outputs a transformed counts matrix using the specified transformation.
}
