% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoRdseq.R
\name{recoRdSeqAnalysis}
\alias{recoRdSeqAnalysis}
\title{Perform Record-seq secondary analysis}
\usage{
recoRdSeqAnalysis(
  countsMatrix,
  designMatrix,
  outPath,
  totalCountsFile = NULL,
  designFormula = NA,
  nGenes = 50,
  K = 2,
  minCountsPerSample = 500,
  geneBoxAndWhiskerPlots = TRUE,
  transformation = "rlog",
  clustering = FALSE,
  PCAplots = TRUE,
  vennDiagrams = TRUE
)
}
\arguments{
\item{countsMatrix}{path to featureCounts output counts matrix}

\item{designMatrix}{path to design matrix}

\item{outPath}{path for saving output files}
}
\description{
This function implements the complete Record-seq analysis pipeline. It takes paths of files generated by 
Record-seq primary analysis as input, and generates all output files at desired output path.
}
