% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoRdseq.R
\name{spacerStats}
\alias{spacerStats}
\title{Create descriptive plots for spacers counts and sequences}
\usage{
spacerStats(
  summaryStats = NULL,
  SpacerInfoFiles_Path = NULL,
  outPath = "/path/for/output",
  designMatrix = "path/to/designMatrix"
)
}
\arguments{
\item{summaryStats}{path to summaryStats file generated by primary analysis pipeline}

\item{SpacerInfoFiles_Path}{path to directory containing spacer sequence info files generated by primary analysis pipeline}

\item{outPath}{output path}

\item{designMatrix}{path to design matrix file}
}
\description{
This function takes in a (optional) spacer stats (summaryStats) file, design matrix file, (optional) spacer sequence info files, 
and generates descriptive plots at a specified output path.
}
