% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoRdseq.R
\name{zscorestandardize}
\alias{zscorestandardize}
\title{Return z-score standardized values}
\usage{
zscorestandardize(x)
}
\arguments{
\item{x}{input vector or matrix}
}
\value{
z-score standardized input
}
\description{
This function takes in a numerical vector or matrix. It calculates and returns the z-score standardized value of each element.
}
